MP_SIL: Convert all short/int/long/longlong to int16_t/int32_t/int32_t/int64_t for cross-platform compatibility
git-svn-id: https://gentlenav.googlecode.com/svn/branches/MatrixPilot_SIL@2042 cf4e4d4a-e611-11dd-8fb1-9960f2a117f8

+++ MatrixPilot/behavior.c
@@ -21,11 +21,11 @@
 
 #include "defines.h"
 
-int current_orientation ;
+int16_t current_orientation ;
 union bfbts_word desired_behavior ;
 
-int cyclesUntilStartTriggerAction = 0 ;
-int cyclesUntilStopTriggerAction = 0 ;
+int16_t cyclesUntilStartTriggerAction = 0 ;
+int16_t cyclesUntilStopTriggerAction = 0 ;
 boolean currentTriggerActionValue = 0 ;
 
 void triggerActionSetValue( boolean newValue ) ;
@@ -48,7 +48,7 @@ void init_behavior( void )
 }
 
 
-void setBehavior(int newBehavior)
+void setBehavior(int16_t newBehavior)
 {
 	desired_behavior.W = newBehavior ;
 	
@@ -160,7 +160,7 @@ void updateTriggerAction( void )
 		{
 			triggerActionSetValue( TRIGGER_ACTION == TRIGGER_PULSE_HIGH ) ;
 			
-			cyclesUntilStopTriggerAction = TRIGGER_PULSE_DURATION / (long)25 ;
+			cyclesUntilStopTriggerAction = TRIGGER_PULSE_DURATION / (int32_t)25 ;
 			cyclesUntilStartTriggerAction = 0 ;
 		}
 		else if ( TRIGGER_ACTION == TRIGGER_TOGGLE )
@@ -174,8 +174,8 @@ void updateTriggerAction( void )
 		{
 			triggerActionSetValue( TRIGGER_ACTION == TRIGGER_PULSE_HIGH ) ;
 			
-			cyclesUntilStopTriggerAction = TRIGGER_PULSE_DURATION / (long)25 ;
-			cyclesUntilStartTriggerAction = TRIGGER_REPEAT_PERIOD / (long)25 ;
+			cyclesUntilStopTriggerAction = TRIGGER_PULSE_DURATION / (int32_t)25 ;
+			cyclesUntilStartTriggerAction = TRIGGER_REPEAT_PERIOD / (int32_t)25 ;
 		}
 	}
 	else if ( cyclesUntilStartTriggerAction > 0 )
